+++ direct_cli/_vendor/tapi_yandex_direct/__init__.py
@@ -1,7 +1,6 @@
-
-__author__ = 'Pavel Maksimov'
-__email__ = 'vur21@ya.ru'
-__version__ = '2026.5.29'
+__author__ = "Pavel Maksimov"
+__email__ = "vur21@ya.ru"
+__version__ = "2026.5.29"
 
 
 from .resource_mapping import *